requirements.txt
greenlet==3.2.4
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
typing_extensions==4.14.1
